remove CavedbLatLonFormField; all lat/lons need to be entered in decimal
This is due to Django thinking that this field is always changing. I
suspect it's due to the model is a DecimalField, but the admin is a
CharField. Changing the admin to a DecimalField does not allow entering
non-decimal numbers. Let's just yank out all of this code for
simplicity.